Fix missing fixtures
